feat: complete owner product management, i18n consistency, and test coverage
- Add owner product CRUD (list/add/edit/delete), brand management, and
  security hardening (timing-safe owner login, lockout after repeated
  failures, env-driven production security settings, CI workflow)
- Fix EN/UK localization gaps: owner dashboard, auth screens, form/view
  messages, order/appointment statuses, and currency were previously
  hardcoded in a single language; rebuilt locale/uk/LC_MESSAGES/django.po
  to match actual template strings
- Add django-modeltranslation for DB content (brand/product descriptions,
  service name/description, specialist bio) that gettext can't reach,
  with real English translations of the existing Ukrainian content;
  specialist's real name intentionally left untranslated
- Expand test suite from 30 to 61 tests: reviews app (previously untested),
  cart edge cases, wishlist isolation, catalog/owner CRUD access control,
  owner dashboard view
- Remove unused accounts/login.html and register.html (superseded by
  auth.html); add real screenshots to README

+++ cart/cart.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from catalog.models import Product
 
 CART_SESSION_KEY = "cart"
@@ -51,7 +53,7 @@ class Cart:
         for item in self.cart.values():
             yield {
                 **item,
-                "total": float(item["price"]) * item["quantity"],
+                "total": Decimal(item["price"]) * item["quantity"],
             }
 
     def __len__(self):
@@ -59,4 +61,7 @@ class Cart:
 
     @property
     def total_price(self):
-        return sum(float(item["price"]) * item["quantity"] for item in self.cart.values())
+        return sum(
+            Decimal(item["price"]) * item["quantity"]
+            for item in self.cart.values()
+        )
